requirements.txt
browser-use
langchain-openai
langchain-anthropic
langchain-google-genai
openai
anthropic
google-genai
python-dotenv
rich
